+++ src/App.js
@@ -2,19 +2,21 @@
 import React, { useState } from 'react';
 import FormRegistro from './proyectofinal/Registro';
 import FormLog from './proyectofinal/Login';
-import HomePage from './proyectofinal/menu';
+import Menu from './proyectofinal/menu';
 import './proyectofinal/main.css';
 
 function App() {
   const [usuarios, setUsuarios] = useState([]);
   const [mostrarFormRegistro, setMostrarFormRegistro] = useState(false);
   const [isLoggedIn, setIsLoggedIn] = useState(false)
+ 
 
   const handleRegistro = (nuevoUsuario) => {
     setUsuarios([...usuarios, nuevoUsuario]);
     alert('¡Registro exitoso!');
   };
   const handleInicioSesion = (credenciales) => {
+    const usuarios = [{nombre:'123', correo:"123@123.com", celular: "123", contraseña: "123"} ];
     const usuarioEncontrado = usuarios.find(
       (usuario) => usuario.correo === credenciales.correo && usuario.contraseña === credenciales.contraseña
     );
@@ -25,10 +27,11 @@ function App() {
       alert('Correo electrónico o contraseña incorrectos');
     }
   };
+  
   return (
     <div>
       {isLoggedIn ? ( 
-        <HomePage />
+        <Menu />
       ) : (
       <div>
         <button class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded" onClick={() => setMostrarFormRegistro(true)}>Registrarse</button>
